add 'like' and 'views'
